Team branch

+++ src/components/manageStatus.vue
@@ -1,18 +1,15 @@
 <script setup>
-import { ref, onMounted } from "vue";
+import { ref, onMounted, watch } from "vue";
 import { useRouter } from "vue-router";
 import { useToast } from "vue-toastification";
+import { computed } from "vue";
+
 const toast = useToast();
 const router = useRouter();
 let statuses = ref([]);
 const showDeleteModal = ref(false);
 let selectedDeletedStatus = ref(null);
-
-const openDeleteModal = (status) => {
-  selectedDeletedStatus.value = status;
-  showDeleteModal.value = true;
-};
-
+let selectedStatusTaskCount = ref(0);
 let showTransferModal = ref(false);
 let newStatus = ref(null);
 
@@ -28,19 +25,38 @@ const gotoEditStatus = (status) => {
   router.push({ path: `/task/status/${status.id}` });
 };
 
+const openDeleteModal = async (status) => {
+  try {
+    if (status) {
+      selectedDeletedStatus.value = status;
+      const tasks = await fetchTasksByStatus(status.id);
+      const isStatusInUse = tasks.length > 0;
+      // console.log("tasks:", tasks);
+      if (isStatusInUse) {
+        showTransferModal.value = true;
+      } else {
+        showDeleteModal.value = true;
+      }
+    }
+  } catch (error) {
+    console.error("Error:", error);
+  }
+};
+
 const confirmDelete = async () => {
   try {
-    const tasks = await fetchTasksByStatus(selectedDeletedStatus.value.id);
-    const isStatusInUse = tasks.length > 0;
-    console.log(fetchTasksByStatus(selectedDeletedStatus.value.id));
-
-    if (isStatusInUse) {
-      showTransferModal.value = true;
-    } else {
-      await deleteStatus(selectedDeletedStatus.value.id);
-      showDeleteModal.value = false;
-      showTransferModal.value = false;
-      await fetchStatuses();
+    if (selectedDeletedStatus.value) {
+      const tasks = await fetchTasksByStatus(selectedDeletedStatus.value.id);
+      const isStatusInUse = tasks.length > 0;
+      if (isStatusInUse) {
+        showTransferModal.value = true;
+        showDeleteModal.value = false;
+      } else {
+        await deleteStatus(selectedDeletedStatus.value.id);
+        showDeleteModal.value = false;
+        showTransferModal.value = false;
+        await fetchStatuses();
+      }
     }
   } catch (error) {
     console.error("Error:", error);
@@ -63,6 +79,9 @@ const deleteStatus = async (statusId, newStatusId) => {
     } else {
       toast.error("Failed to delete status");
     }
+    setTimeout(() => {
+      location.reload();
+    }, 1500);
     throw new Error("Failed to delete status");
   }
   toast.success("The status has been deleted");
@@ -97,7 +116,9 @@ const fetchTasksByStatus = async (statusId) => {
 
 const fetchStatuses = async () => {
   try {
-    const response = await fetch("http://ip23or3.sit.kmutt.ac.th:8080/v2/statuses");
+    const response = await fetch(
+    "http://ip23or3.sit.kmutt.ac.th:8080/v2/statuses"
+    );
     // const response = await fetch("http://localhost:8080/v2/statuses");
     if (!response.ok) {
       throw new Error("Failed to fetch statuses");
@@ -115,6 +136,24 @@ onMounted(async () => {
   if (noStatus) {
     newStatus.value = noStatus.id;
   }
+});
+
+// count tasks for the selected status
+watch(
+  selectedDeletedStatus,
+  async (newStatus) => {
+    if (newStatus) {
+      const tasks = await fetchTasksByStatus(newStatus.id);
+      selectedStatusTaskCount.value = tasks.length;
+    }
+  },
+  { immediate: true }
+);
+
+const availableStatuses = computed(() => {
+  return statuses.value.filter(
+    (status) => status.id !== selectedDeletedStatus.value?.id
+  );
 });
 </script>
 <template>
@@ -177,13 +216,13 @@ onMounted(async () => {
             <td
               class="text-center px-6 py-4 whitespace-nowrap text-sm font-medium text-green-600 border-b border-gray-200"
             >
-              {{ status.description }}
+              {{ status.description ? "" : "No description provided" }}
             </td>
             <td
               class="text-center px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-800 border-b border-gray-200"
             >
               <button
-                v-if="status.name !== 'No Status'"
+                v-if="status.name !== 'no status'"
                 @click="gotoEditStatus(status)"
                 class="itbkk-button-edit btn btn-outline btn-primary bg-blue-200 btn-md"
               >
@@ -201,7 +240,7 @@ onMounted(async () => {
               class="text-center px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-800 border-b border-gray-200"
             >
               <button
-                v-if="status.name !== 'No Status'"
+                v-if="status.name !== 'no status'"
                 @click="openDeleteModal(status)"
                 class="itbkk-button-delete btn btn-outline btn-danger bg-red-200 btn-md"
               >
@@ -224,7 +263,7 @@ onMounted(async () => {
         class="fixed z-10 inset-0 overflow-y-auto flex items-center justify-center bg-slate-500 bg-opacity-25"
       >
         <div
-          class="bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:max-w-lg sm:w-full"
+          class="bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:max-w-lg sm:w-full border-red-400 border-2"
         >
           <div class="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
             <div class="flex flex-col items-center justify-center text-center">
@@ -273,17 +312,18 @@ onMounted(async () => {
         class="fixed z-10 inset-0 overflow-y-auto flex items-center justify-center bg-slate-500 bg-opacity-25"
       >
         <div
-          class="bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:max-w-lg sm:w-full"
+          class="bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:max-w-lg sm:w-full border-red-400 border-2"
         >
           <div class="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
             <div class="flex flex-col items-center justify-center text-center">
               <h3 class="text-lg leading-6 font-medium text-gray-900">
                 The status "{{ selectedDeletedStatus.name }}" is currently in
-                use by a task. Please select a new status for these tasks.
+                use {{ selectedStatusTaskCount }} task. Please select a new
+                status for these tasks.
               </h3>
-              <select v-model="newStatus" class="mt-4">
+              <select v-model="newStatus" class="mt-4 bg-gray-200 rounded-md">
                 <option
-                  v-for="status in statuses"
+                  v-for="status in availableStatuses"
                   :value="status.id"
                   :key="status.id"
                 >
